LayoutTests/fast/dom/shadow/pseudo-attribute-expected.txt

This test checks pseudo attribute is exposed correctly.

PASS div.pseudo is ''
PASS div.getAttribute('pseudo') is null
PASS div.pseudo is 'foo'
PASS div.getAttribute('pseudo') is 'foo'
PASS div.pseudo is 'bar'
PASS div.getAttribute('pseudo') is 'bar'
PASS div.pseudo is ''
PASS div.getAttribute('pseudo') is null
PASS div.pseudo is 'first-letter'
PASS successfullyParsed is true

TEST COMPLETE
